Move attributes methods to Attributes plugin
