+++ lib/mobility/plugins/attributes.rb
@@ -4,7 +4,8 @@ module Mobility
 =begin
 
 Takes arguments, converts them to strings, and stores in an array +@names+,
-made available with an +attr_reader+.
+made available with an +attr_reader+. Also provides some convenience methods
+for aggregating attributes.
 
 =end
     module Attributes
@@ -28,6 +29,32 @@ made available with an +attr_reader+.
       # @return [String]
       def inspect
         "#<Attributes @names=#{names.join(", ")}>"
+      end
+
+      included_hook do |klass, *, **|
+        klass.extend ClassMethods
+      end
+
+      module ClassMethods
+        # Return all {Mobility::Attributes} module instances from among ancestors
+        # of this model.
+        # @return [Array<Mobility::Attributes>] Attribute modules
+        def mobility_modules
+          ancestors.grep(Attributes)
+        end
+
+        # Return translated attribute names on this model.
+        # @return [Array<String>] Attribute names
+        def mobility_attributes
+          mobility_modules.map(&:names).flatten.uniq
+        end
+
+        # Return true if attribute name is translated on this model.
+        # @param [String, Symbol] Attribute name
+        # @return [Boolean]
+        def mobility_attribute?(name)
+          mobility_attributes.include?(name.to_s)
+        end
       end
     end
